Enhance the meeting copy functionality
