+++ decidim-meetings/app/forms/decidim/meetings/meeting_form.rb
@@ -3,22 +3,15 @@
 module Decidim
   module Meetings
     # This class holds a Form to create/update meetings for Participants and UserGroups.
-    class MeetingForm < Decidim::Form
+    class MeetingForm < ::Decidim::Meetings::BaseMeetingForm
       attribute :title, String
       attribute :description, String
       attribute :location, String
       attribute :location_hints, String
 
-      attribute :address, String
-      attribute :latitude, Float
-      attribute :longitude, Float
-      attribute :start_time, Decidim::Attributes::TimeWithZone
-      attribute :end_time, Decidim::Attributes::TimeWithZone
       attribute :decidim_scope_id, Integer
       attribute :decidim_category_id, Integer
       attribute :user_group_id, Integer
-      attribute :online_meeting_url, String
-      attribute :type_of_meeting, String
       attribute :registration_type, String
       attribute :registrations_enabled, Boolean, default: false
       attribute :registration_url, String
@@ -31,17 +24,11 @@ module Decidim
       validates :description, presence: true
       validates :type_of_meeting, presence: true
       validates :location, presence: true, if: ->(form) { form.in_person_meeting? || form.hybrid_meeting? }
-      validates :address, presence: true, if: ->(form) { form.needs_address? }
-      validates :address, geocoding: true, if: ->(form) { form.has_address? && !form.geocoded? && form.needs_address? }
       validates :online_meeting_url, presence: true, url: true, if: ->(form) { form.online_meeting? || form.hybrid_meeting? }
       validates :registration_type, presence: true
       validates :available_slots, numericality: { greater_than_or_equal_to: 0 }, presence: true, if: ->(form) { form.on_this_platform? }
       validates :registration_terms, presence: true, if: ->(form) { form.on_this_platform? }
       validates :registration_url, presence: true, url: true, if: ->(form) { form.on_different_platform? }
-      validates :start_time, presence: true, date: { before: :end_time }
-      validates :end_time, presence: true, date: { after: :start_time }
-
-      validates :current_component, presence: true
       validates :category, presence: true, if: ->(form) { form.decidim_category_id.present? }
       validates :scope, presence: true, if: ->(form) { form.decidim_scope_id.present? }
       validates :decidim_scope_id, scope_belongs_to_component: true, if: ->(form) { form.decidim_scope_id.present? }
@@ -81,34 +68,6 @@ module Decidim
         return unless current_component
 
         @category ||= categories.find_by(id: decidim_category_id)
-      end
-
-      def geocoding_enabled?
-        Decidim::Map.available?(:geocoding)
-      end
-
-      def has_address?
-        geocoding_enabled? && address.present?
-      end
-
-      def needs_address?
-        in_person_meeting? || hybrid_meeting?
-      end
-
-      def geocoded?
-        latitude.present? && longitude.present?
-      end
-
-      def online_meeting?
-        type_of_meeting == "online"
-      end
-
-      def in_person_meeting?
-        type_of_meeting == "in_person"
-      end
-
-      def hybrid_meeting?
-        type_of_meeting == "hybrid"
       end
 
       def clean_type_of_meeting
